Adapt listeners taking fewer args than the signal
Previously an ArgumentError would be thrown if you attempted to add a function listener to a signal
that took less arguments than the signal dispatched. This wraps the passed in listener function in
a function that trims the argument list length to the length of the listener function and then calls
through.

This allows listeners that only care that an event occurred to be more loosely coupled with the
signal; if they're not using the arguments, they don't need to change if the signal argument length
changes.