Adds conditional listeners.
Adds an "addConditionally" method to ISignal.ISignal#addConditionally ad
ds a "conditional" listener to the signal. If the listener returns a tru
thy value (true, > 0, not null) when it's invoked, the listener is remov
ed. If it returns a falsey value (false, 0, null, void), the listener is
n't removed.

For example, this is useful for a "fire-hose" Signal, whic
h can dispatch multiple types.
>	const firehose:ISignal = service.subscr
ibeToSharedEventStream();
>	const typeAHandler:Function = function(obj:*
):void{
>		if(!(obj is TypeA)) return;
>		//else do something...
>		fire
hose.remove(typeAHandler);
>	}
>	signal.add(typeAHandler);

>	service.su
bscribeToSharedEventStream()
>	.addConditionally(function(obj:*):Boolean
{
>		if(!(obj is TypeA)) return false;
>		//else do whatever and remove
the listener
>		return true;
>	});